requirements.txt

Flask>=2.0
openai>=0.27.0
python-dotenv
